No errors commit
Code is stable after some initial refactoring to MediaLibrary, GUIController, MediaPlayerManager,  and Track. Made Track an abstract class with GenericTrack, MusicSampleTrack, and GuitarTrack concrete classes that refer to this class. Made a TrackFactory class tasked with creating Tracks based on what type that is passed to it. Next step is to develop TrackFactory methods where the AudioSource file gets passed from the MediaLibrary AddTrack method