Playback Mods 2.0
Playback Play(), auto-advance, playback bar all work seamlessly now. Disabled any references to PlayableGraph in Track as it's functionality is redundant to that of AudioSource and was causing issues in the Stop method (PlayableGraph's Stop method is more like a pause). Track methods all reworked to look for AudioSource references and AudioSource references only