MenuComponent Refactored
MenuComponent was refactored and now Pointer events work uniformly for context menus and static menus. Need to do some calibrating of anchors on the context menu prefab or add a tolerance to boundary detection so the context menu works a little better. Need to do some general cleanup as well and then we will merge back to main