Serialization/Deserialization methods added to TrackLibaryController
Also a PopulateOnLoad and "TracksDeserialized" observation type were added to UITrackListDisplayManager. Default MyDocuments path defaults to OneDrive/Documents/... which is kind of annoying but we'll live with that for now. No build errors but program doesn't seem to actually be loading anything so will need to do some debugging. Serialization/deserialization probably needs to be fine-tuned