MultiFile Upload is implemented
Need to do some more debugging on the Remove feature as when I updated the AddTrackCommand and TrackLibraryController to allow for multi-file upload I made similar changes to RemoveTrackCommand and TrackLibraryController to allow for mass file removal. however, these changes are now causing bugs with singular removal so will need to see why this is. I'm also noticing that my scroll view does not work properly now that I have enough tracks in the library to warrant scrolling