Some light changes to UISampleDisplayManager, but still have bugs
State change won't trigger text change and I think the issue is in how HandleSelectedTrackSamplesAndSampledBysConfiguration() is setup. I may need to move this method to GeniusWebClientController but not sure (that class is getting crowded so may want to break this into seperate classes for URL and sample searching). May also need to configure an "OnGeniusSampleSearchcomplete" observation and actually reference HandleFetchGeniusSampleCompletion() somewhere in Genius controller