Refactored UISampleDisplayManager and GeniusWebController to improve functionality
Works pretty good, but there are still some bugs when selecting a track with NotFound data and then re-selecting a different track with Found data. Clicking again on the new track will update to show the correct data. Not sure why this issue is happening. May need to do some in depth debugging or create some unit tests or something. Breaking up GeniusWebClientController might be helpful as well